rename 'cannot obtain cover' to 'cannot apply'
