+++ e2e-tests/constants/routes/quote.js
@@ -18,7 +18,7 @@ const QUOTE_ROUTES = {
 
   CHECK_YOUR_ANSWERS: '/quote/check-your-answers',
 
-  CANNOT_OBTAIN_COVER: '/quote/cannot-obtain-cover',
+  CANNOT_APPLY: '/quote/cannot-apply',
 
   GET_A_QUOTE_BY_EMAIL: '/quote/get-quote-by-email',
