+++ e2e-tests/cypress/e2e/journeys/quote/buyer-country/buyer-country-unsupported-country.spec.js
@@ -27,7 +27,7 @@ context('Buyer country page - as an exporter, I want to check if UKEF issue expo
   });
 
   it('redirects to `cannot obtain cover` exit page', () => {
-    cy.url().should('include', ROUTES.QUOTE.CANNOT_OBTAIN_COVER);
+    cy.url().should('include', ROUTES.QUOTE.CANNOT_APPLY);
   });
 
   it('renders a back link with correct url', () => {
