+++ src/ui/server/constants/routes/quote.ts
@@ -18,7 +18,7 @@ export const QUOTE_ROUTES = {
 
   CHECK_YOUR_ANSWERS: '/quote/check-your-answers',
 
-  CANNOT_OBTAIN_COVER: '/quote/cannot-obtain-cover',
+  CANNOT_APPLY: '/quote/cannot-apply',
 
   GET_A_QUOTE_BY_EMAIL: '/quote/get-quote-by-email',
