+++ src/ui/server/constants/templates/quote.ts
@@ -4,7 +4,7 @@ export const QUOTE_TEMPLATES = {
   POLICY_TYPE: 'quote/policy-type.njk',
   TELL_US_ABOUT_YOUR_POLICY: 'quote/tell-us-about-your-policy.njk',
   CHECK_YOUR_ANSWERS: 'quote/check-your-answers.njk',
-  CANNOT_OBTAIN_COVER: 'quote/cannot-obtain-cover.njk',
+  CANNOT_APPLY: 'quote/cannot-apply.njk',
   GET_A_QUOTE_BY_EMAIL: 'quote/get-a-quote-by-email.njk',
   YOUR_QUOTE: 'quote/your-quote.njk',
   NEED_TO_START_AGAIN: 'quote/need-to-start-again.njk',
